test: split domain-coverage.test.ts to satisfy 1200-line CI hard limit
CI's check-file-sizes.sh failed because the file grew to 1351 lines after
adding 18 condition-count tests + the disjoint-slots regression tests.

Split:
- domain-coverage.test.ts (835 lines): unit tests for individual exports
- domain-coverage-integration.test.ts (527 lines): functional integration scenarios
  that exercise the directional inner-loop construction across multiple helpers

Both files share imports from domain-coverage-utils. All 81 tests still pass.
