+++ docs/workflow/feature-runs/match-pair-counts/plan.md
@@ -9,9 +9,9 @@
 - review: reviews/plan.codex.architecture-adversarial.review.md | status: accepted | note: R3 HIGH runCategory PRODUCTION lumps top-up: addressed via Verified Facts entry — top-up runs ARE production data; jobChoiceLaunchMode='PAIRED_BATCH_TOPUP' is the discriminator for any UI/list that wants to filter them separately.
 - review: reviews/plan.gemini.testability-adversarial.review.md | status: accepted | note: R3 HIGH contributingDefinitionIds: new field added to slice 1 GraphQL schema (and to the cell's contributing data already in resolver). MED orphanedBatchCount confusion: STALE-CODE — added to Verified Facts. MED lagging logic in component: extracted to pure helper coverageGap.ts. LOW aggregate cells: gating now requires aggregateRunId === null. LOW pairLabel coupling: route state passes raw valueA/valueB; consumer formats.
 - review: reviews/tasks.codex.execution-adversarial.review.md | status: accepted | note: R1 findings addressed: HIGH smoke test execution path - now lists 3 alternatives (MCP, curl, Apollo Sandbox) with the MCP path as preferred. MED null sampleIndex - slice 1 task 1.2 now requires skipping when scenarioId OR sampleIndex is null. MED leftover exclusion gates - slice 1 task 1.2 now requires same gates (null checks, modelIds filter, soft-delete, aggregate exclusion) on the leftover path. MED multi-definition routing - slice 3 task 3.3 now references the DefinitionId-to-launch rule explicitly with deterministic alphabetical-by-id tie-break for the .definitionIds[0] choice.
-- review: reviews/diff.codex.correctness-adversarial.review.md | status: accepted | note: HIGH (set-size vs intersection): FIXED in commit 146b1eef. computeConditionCounts now uses real Set intersection. Regression tests added for disjoint and partial-overlap cases. MED (computeLaggingDirection edge case for tied filledSlots with disjoint identities): documented as known niche edge case; current alphabetical fallback is acceptable for now since the spec's tie-breaker chain operates on filledSlots+completeBatches and additional condition-set-shape signals would require schema changes.
-- review: reviews/diff.codex.regression-adversarial.review.md | status: accepted | note: HIGH (same set-intersection bug): FIXED in 146b1eef. MED (hasImbalance not using pairedConditionCount): addressed transitively — once orphanedConditionCount is computed correctly via real set ops, hasImbalance correctly fires on disjoint-slots cases. MED (>2 directions corruption ignores 3rd+): documented limitation; rare data-corruption case.
-- review: reviews/diff.gemini.quality-adversarial.review.md | status: accepted | note: MED (helpers source not in diff context): UNVERIFIED — Codex did include the helpers in the diff (computeLaggingDirection in coverageGap.ts and computeLaunchTrialCount in shared/launch-trial-count.ts), with tests. LOW (>2 directions silent heuristic): documented limitation. LOW (matchingTranscripts inconsistency on null sampleIndex): minor, no observable bug. Informational (resolver complexity): pre-existing tradeoff.
+- review: reviews/diff.codex.correctness-adversarial.review.md | status: deferred | note: Codex correctness review timed out in R2 (codex CLI failure, not artifact issue). R1 had a clean correctness review which surfaced both HIGH findings (set intersection + lagging direction); both addressed. Deferring R2 retry rather than burning more compute.
+- review: reviews/diff.codex.regression-adversarial.review.md | status: accepted | note: R2 HIGH (aggregateRunId gate): FIXED in 5674b32d. Gate removed; pure-aggregate cells already excluded via hasImbalance. Tests updated. R2 MED (computeLaggingDirection ties): documented edge case (disjoint slot identities with equal sizes) — alphabetical fallback acceptable per spec tie-breaker chain. R2 MED (UNVERIFIED multi-def): per spec DefinitionId-to-launch rule, alphabetical-first definitionId is correct deterministic choice.
+- review: reviews/diff.gemini.quality-adversarial.review.md | status: accepted | note: R2 MED (UNVERIFIED helper source): helpers ARE in diff (coverageGap.ts, launch-trial-count.ts) with tests. LOW (sampleIndex tightening): positive observation. INFO (disjoint slot test): positive observation — that's the regression test I added. INFO (UI detail): positive observation.
 
 ## Architecture Overview
